// metis-web/packages/api/src/generated/SessionVersionRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActorRef } from "./ActorRef";
import type { Session } from "./Session";
import type { SessionId } from "./SessionId";

export type SessionVersionRecord = { session_id: SessionId, version: bigint, timestamp: string, session: Session, actor?: ActorRef | null, };